Fully implement dynamic post editting
Posts now dynamically update when a user clicks the 'edit' button and
also dynamically update when the user submits a new post as well. This
now works on the home (all) posts page as well as when a user is viewing
their own profile.

In addition, the most up-to-date versions of posts will show when a user
is moving between pages of posts as well.